refactor: Public API 추가 및 경로 수정
