+++ apps/web/entities/post/index.ts
@@ -1,4 +1,10 @@
 export { createPost, getPostById, getPosts } from './api/post-api';
 export { ERROR_MESSAGES } from './lib/error-messages';
 export { validateFormField } from './lib/form-validation';
-export type { CreatePostDTO, Post, UpdatePostDTO } from './model/post-types';
+export type {
+  CreatePostDTO,
+  GetPostsCursor,
+  Post,
+  UpdatePostDTO,
+} from './model/post-types';
+export { mapPostToViewModel } from './model/post-view-model';
