+++ apps/web/entities/post/model/post-view-model.ts
@@ -1,5 +1,5 @@
-import { formatToLocaleDate } from '@/shared/lib/format-date';
-import { Post } from '@/shared/types/post-types';
+import { Post } from '@/entities/post';
+import { formatToLocaleDate } from '@/shared/lib';
 
 export const mapPostToViewModel = (post: Post) => ({
   ...post,
